// src/api/generated/models/updateMemberRoleDtoRole.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Flowforge API
 * The Flowforge workflow orchestration engine API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * New role for the member (ADMIN or MEMBER)
 */
export type UpdateMemberRoleDtoRole = typeof UpdateMemberRoleDtoRole[keyof typeof UpdateMemberRoleDtoRole];


export const UpdateMemberRoleDtoRole = {
  ADMIN: 'ADMIN',
  MEMBER: 'MEMBER',
} as const;